refactor - exclude styles data
