+++ src/types/info.ts
@@ -8,7 +8,8 @@ export type BasicUserInfo = {
 };
 
 export type UserInfo = {
-  styles: number[];
+  // TODO: 서비스 고도화 전까지 styles 데이터 주석
+  // styles: number[];
   gender: string;
   height: string;
   bodyShape: string;
