+++ src/types/storeType/info.ts
@@ -1,7 +1,8 @@
 import { UserInfo } from '#types/info';
 
 export interface InfoState {
-  styles: number[];
+  // TODO: 서비스 고도화 전까지 styles 데이터 주석
+  // styles: number[];
   gender: string;
   height: number;
   bodyShape: string;
